исправления с формой авторизации, токен с сервера, readonly для поля ввода
