+++ api.js
@@ -1,7 +1,6 @@
-const token = "bgc0b8awbwas6g5g5k5o5s5w606g37w3cc3bo3b83k39s3co3c83c03ck";
-localStorage.setItem("token", token);
+const token = localStorage.getItem("token");
 
-export function getTodos(token) {
+export function getTodos() {
     return fetch("https://wedev-api.sky.pro/api/v2/karina-korneva/comments", {
         headers: {
             Authorization: `Bearer ${token}`
