+++ index_hw-9.html
@@ -15,7 +15,8 @@
     <div class="add-authorization" id="auth-message">Чтобы добавить комментарий, <span class="link-login" id="login-link">авторизуйтесь</span>.</div>
     <div class="add-form">
         <input
-            type="text"
+            type="text" 
+            readonly 
             class="add-form-name"
             placeholder="Введите ваше имя"
         />
