+++ styles_hw-9.css
@@ -103,7 +103,6 @@ body {
   .add-form {
     padding: 20px;
     margin-top: 48px;
-    display: flex;
     flex-direction: column;
     display: none;
   }
